--- linkify-plus-plus.user.js
@@ -1,105 +1,0 @@
-// ==UserScript==
-// @name Linkify Plus Plus
-// @version 8.2.2
-// @description Based on Linkify Plus. Turn plain text URLs into links.
-// @license BSD-3-Clause
-// @homepageURL https://github.com/eight04/linkify-plus-plus
-// @supportURL https://github.com/eight04/linkify-plus-plus/issues
-// @namespace eight04.blogspot.com
-// @include *
-// @exclude https://www.google.*/search*
-// @exclude https://www.google.*/webhp*
-// @exclude https://music.google.com/*
-// @exclude https://mail.google.com/*
-// @exclude https://docs.google.com/*
-// @exclude https://encrypted.google.com/*
-// @exclude http://mxr.mozilla.org/*
-// @exclude http://w3c*.github.io/*
-// @require https://greasyfork.org/scripts/7212-gm-config-eight-s-version/code/GM_config%20(eight's%20version).js?version=156587
-// @require https://greasyfork.org/scripts/27630-linkify-plus-plus-core/code/linkify-plus-plus-core.js?version=213494
-// @grant GM_addStyle
-// @grant GM_registerMenuCommand
-// @grant GM_getValue
-// @grant GM_setValue
-// @grant unsafeWindow
-// @compatible firefox Tampermonkey latest
-// @compatible chrome Tampermonkey latest
-// ==/UserScript==
-
-/* globals GM_config */
-
-(function(){
-
-
-
-
-// Program init
-GM_addStyle(".linkifyplus img { max-width: 90%; }");
-const linkifyProcess = createLinkifyProcess(createOptions());
-const observer = new MutationObserver(function(mutations){
-	// Filter out mutations generated by LPP
-	var lastRecord = mutations[mutations.length - 1],
-		nodes = lastRecord.addedNodes,
-		i;
-
-	if (nodes.length >= 2) {
-		for (i = 0; i < 2; i++) {
-			if (nodes[i].className == "linkifyplus") {
-				return;
-			}
-		}
-	}
-
-  for (var record of mutations) {
-    if (record.addedNodes.length) {
-      if (!linkifyProcess.process(record.target)) {
-        // it's full
-        break;
-      }
-    }
-  }
-});
-
-function init() {
-  observer.observe(document.body, {
-    childList: true,
-    subtree: true
-  });
-  linkifyProcess.process(document.body);
-}
-
-function prepare() {
-  // wait till everything is ready
-  return Promise.all([prepareApp(), prepareBody()]);
-  
-  function prepareApp() {
-    const appRoot = document.querySelector("[data-server-rendered]");
-    if (!appRoot) {
-      return;
-    }
-    return new Promise(resolve => {
-      const onChange = () => {
-        if (!appRoot.hasAttribute("data-server-rendered")) {
-          resolve();
-          observer.disconnect();
-        }
-      };
-      const observer = new MutationObserver(onChange);
-      observer.observe(appRoot, {attributes: true});
-    });
-  }
-  
-  function prepareBody() {
-    if (document.body) {
-      return;
-    }
-    return new Promise(resolve => {
-      // https://github.com/Tampermonkey/tampermonkey/issues/485
-      document.addEventListener("DOMContentLoaded", resolve, {once: true});
-    });
-  }
-}
-
-prepare().then(init);
-
-})();
