+++ src/lib/linkify-plus-plus.js
@@ -43,16 +43,18 @@ function validRoot(node, validator) {
   return isValid;
 }
 
-function createValidator({selector, skipSelector}) {
+function createValidator({includeElement, excludeElement}) {
   return function(node) {
     if (node.isContentEditable) {
       return false;
     }
-    if (selector && node.matches && node.matches(selector)) {
-      return true;
-    }
-    if (skipSelector && node.matches && node.matches(skipSelector)) {
-      return false;
+    if (node.matches) {
+      if (includeElement && node.matches(includeElement)) {
+        return true;
+      }
+      if (excludeElement && node.matches(excludeElement)) {
+        return false;
+      }
     }
     return true;
   };
@@ -136,8 +138,8 @@ function createLinkifyProcess(options) {
     linkify(root, options)
       .then(() => {
         var p = Promise.resolve();
-        if (options.selector) {
-          for (var node of root.querySelectorAll(options.selector)) {
+        if (options.includeElement) {
+          for (var node of root.querySelectorAll(options.includeElement)) {
             p = p.then(linkify.bind(null, node, options));
           }
         }
@@ -165,33 +167,17 @@ function createOptions(pref) {
     options.onlink = options.embedImageExcludeElement ? onlink : null;
   }
   
-  function setup() {
-    }, function() {
-      Object.assign(options, GM_config.get());
-      if (options.selector && !selectorTest(options.selector)) {
-        options.selector = null;
-      }
-      if (options.skipSelector && !selectorTest(options.skipSelector)) {
-        options.skipSelector = null;
-      }
-      if (options.customRules) {
-        options.customRules = createList(options.customRules);
-      }
-      options.validator = createValidator(options);
-      options.fuzzyIp = options.ip;
-      options.embedImage = options.image;
-    });
-  }
-  
   function onlink({link, range, content}) {
-    if (link.childNodes[0].nodeName != "IMG") return;
+    if (link.childNodes[0].nodeName != "IMG" || !options.embedImageExcludeElement) {
+      return;
+    }
     
     var parent = range.startContainer;
     // it might be a text node
     if (!parent.closest) {
       parent = parent.parentNode;
     }
-    if (!parent.closest(options.imageSkipSelector)) return;
+    if (!parent.closest(options.embedImageExcludeElement)) return;
     // remove image
     link.innerHTML = "";
     link.appendChild(content);
@@ -200,13 +186,80 @@ function createOptions(pref) {
 
 function startLinkifyPlusPlus(pref) {
   // Limit contentType to "text/plain" or "text/html"
-  if (document.contentType != undefined && document.contentType != "text/plain" && document.contentType != "text/html") {
+  if (
+    document.contentType !== undefined &&
+    document.contentType != "text/plain" &&
+    document.contentType != "text/html"
+  ) {
     return;
   }
   
   const {linkify, UrlMatcher, INVALID_TAGS} = linkifyPlusPlusCore;
   const BUFFER_SIZE = 100;
   const linkifyProcess = createLinkifyProcess(createOptions(pref));  
+  const observer = new MutationObserver(function(mutations){
+    // Filter out mutations generated by LPP
+    var lastRecord = mutations[mutations.length - 1],
+      nodes = lastRecord.addedNodes,
+      i;
+
+    if (nodes.length >= 2) {
+      for (i = 0; i < 2; i++) {
+        if (nodes[i].className == "linkifyplus") {
+          return;
+        }
+      }
+    }
+
+    for (var record of mutations) {
+      if (record.addedNodes.length) {
+        if (!linkifyProcess.process(record.target)) {
+          // it's full
+          break;
+        }
+      }
+    }
+  });
+  
+  prepareDocument().then(() => {
+    observer.observe(document.body, {
+      childList: true,
+      subtree: true
+    });
+    linkifyProcess.process(document.body);    
+  });
+}
+
+function prepareDocument() {
+  // wait till everything is ready
+  return prepareBody().then(prepareApp);
+  
+  function prepareApp() {
+    const appRoot = document.querySelector("[data-server-rendered]");
+    if (!appRoot) {
+      return;
+    }
+    return new Promise(resolve => {
+      const onChange = () => {
+        if (!appRoot.hasAttribute("data-server-rendered")) {
+          resolve();
+          observer.disconnect();
+        }
+      };
+      const observer = new MutationObserver(onChange);
+      observer.observe(appRoot, {attributes: true});
+    });
+  }
+  
+  function prepareBody() {
+    if (document.readyState !== "loading") {
+      return;
+    }
+    return new Promise(resolve => {
+      // https://github.com/Tampermonkey/tampermonkey/issues/485
+      document.addEventListener("DOMContentLoaded", resolve, {once: true});
+    });
+  }
 }
 
 module.exports = {startLinkifyPlusPlus};
